+++ comics/comics/theidlestate.py
@@ -7,22 +7,11 @@ class ComicData(ComicDataBase):
     language = 'en'
     url = 'http://www.theidlestate.com/'
     start_date = '2011-07-18'
+    end_date = '2012-07-05'
+    active = False
     rights = 'Nick Wright'
 
 
 class Crawler(CrawlerBase):
-    history_capable_days = 40
-    time_zone = 'US/Mountain'
-
     def crawl(self, pub_date):
-        feed = self.parse_feed('http://www.theidlestate.com/?feed=rss2')
-        for entry in feed.for_date(pub_date):
-            if 'Peached' not in entry.tags:
-                continue
-            url = entry.content0.src('img[src*="/wp-content/webcomic/"]')
-            if not url:
-                continue
-            url = url.replace('/thumbs', '')
-            url = url.replace('-medium', '')
-            title = entry.title
-            return CrawlerImage(url, title)
+        pass  # Comic no longer published